requirements.txt
crewai>=0.70.0
crewai-tools>=0.12.0
langchain>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
